Move javasrc statement creation into separate files
