+++ joern-cli/frontends/javasrc2cpg/src/main/scala/io/joern/javasrc2cpg/astcreation/statements/AstForSimpleStatementsCreator.scala
@@ -1,0 +1,304 @@
+package io.joern.javasrc2cpg.astcreation.statements
+
+import io.joern.javasrc2cpg.astcreation.AstCreator
+import com.github.javaparser.ast.stmt.BlockStmt
+import io.joern.x2cpg.Ast
+import io.shiftleft.codepropertygraph.generated.nodes.NewBlock
+import scala.jdk.CollectionConverters.*
+import com.github.javaparser.ast.stmt.ExplicitConstructorInvocationStmt
+import io.joern.javasrc2cpg.typesolvers.TypeInfoCalculator.TypeConstants
+import io.joern.javasrc2cpg.util.NameConstants
+import io.shiftleft.codepropertygraph.generated.EdgeTypes
+import io.joern.x2cpg.utils.NodeBuilders.newIdentifierNode
+import com.github.javaparser.ast.stmt.AssertStmt
+import io.shiftleft.codepropertygraph.generated.nodes.NewCall
+import io.shiftleft.codepropertygraph.generated.DispatchTypes
+import io.joern.javasrc2cpg.astcreation.ExpectedType
+import com.github.javaparser.ast.stmt.BreakStmt
+import io.shiftleft.codepropertygraph.generated.nodes.NewControlStructure
+import com.github.javaparser.ast.stmt.ContinueStmt
+import io.shiftleft.codepropertygraph.generated.ControlStructureTypes
+import com.github.javaparser.ast.stmt.DoStmt
+import com.github.javaparser.ast.stmt.WhileStmt
+import com.github.javaparser.ast.stmt.IfStmt
+import com.github.javaparser.ast.stmt.Statement
+import com.github.javaparser.ast.stmt.SwitchStmt
+import com.github.javaparser.ast.stmt.SynchronizedStmt
+import com.github.javaparser.ast.stmt.SwitchEntry
+import io.shiftleft.codepropertygraph.generated.nodes.NewJumpTarget
+import com.github.javaparser.ast.stmt.ReturnStmt
+import io.shiftleft.codepropertygraph.generated.nodes.NewReturn
+import com.github.javaparser.ast.stmt.LabeledStmt
+import com.github.javaparser.ast.stmt.ThrowStmt
+import com.github.javaparser.ast.stmt.CatchClause
+import com.github.javaparser.ast.stmt.TryStmt
+import io.joern.x2cpg.utils.NodeBuilders.newModifierNode
+
+import scala.jdk.OptionConverters.RichOptional
+
+trait AstForSimpleStatementsCreator { this: AstCreator =>
+  def astForBlockStatement(stmt: BlockStmt, codeStr: String = "<empty>", prefixAsts: Seq[Ast] = Seq.empty): Ast = {
+    val block = NewBlock()
+      .code(codeStr)
+      .lineNumber(line(stmt))
+      .columnNumber(column(stmt))
+
+    scope.pushBlockScope()
+
+    val stmtAsts = stmt.getStatements.asScala.flatMap(astsForStatement)
+
+    scope.popScope()
+    Ast(block)
+      .withChildren(prefixAsts)
+      .withChildren(stmtAsts)
+  }
+
+  private[statements] def astForExplicitConstructorInvocation(stmt: ExplicitConstructorInvocationStmt): Ast = {
+    val maybeResolved = tryWithSafeStackOverflow(stmt.resolve())
+    val args          = argAstsForCall(stmt, maybeResolved, stmt.getArguments)
+    val argTypes      = argumentTypesForMethodLike(maybeResolved)
+
+    val typeFullName = maybeResolved.toOption
+      .map(_.declaringType())
+      .flatMap(typeInfoCalc.fullName)
+
+    val callRoot = initNode(
+      typeFullName.orElse(Some(TypeConstants.Any)),
+      argTypes,
+      args.size,
+      stmt.toString,
+      line(stmt),
+      column(stmt)
+    )
+
+    val thisNode = newIdentifierNode(NameConstants.This, typeFullName.getOrElse(TypeConstants.Any))
+    scope.lookupVariable(NameConstants.This).variableNode.foreach { thisParam =>
+      diffGraph.addEdge(thisNode, thisParam, EdgeTypes.REF)
+    }
+    val thisAst = Ast(thisNode)
+
+    callAst(callRoot, args, Some(thisAst))
+  }
+
+  private[statements] def astForAssertStatement(stmt: AssertStmt): Ast = {
+    val callNode = NewCall()
+      .name("assert")
+      .methodFullName("assert")
+      .dispatchType(DispatchTypes.STATIC_DISPATCH)
+      .code(stmt.toString)
+      .lineNumber(line(stmt))
+      .columnNumber(column(stmt))
+
+    val args = astsForExpression(stmt.getCheck, ExpectedType.Boolean)
+    callAst(callNode, args)
+  }
+
+  private[statements] def astForBreakStatement(stmt: BreakStmt): Ast = {
+    val node = NewControlStructure()
+      .controlStructureType(ControlStructureTypes.BREAK)
+      .lineNumber(line(stmt))
+      .columnNumber(column(stmt))
+      .code(stmt.toString)
+    Ast(node)
+  }
+
+  private[statements] def astForContinueStatement(stmt: ContinueStmt): Ast = {
+    val node = NewControlStructure()
+      .controlStructureType(ControlStructureTypes.CONTINUE)
+      .lineNumber(line(stmt))
+      .columnNumber(column(stmt))
+      .code(stmt.toString)
+    Ast(node)
+  }
+
+  private[statements] def astForDo(stmt: DoStmt): Ast = {
+    val conditionAst = astsForExpression(stmt.getCondition, ExpectedType.Boolean).headOption
+    val stmtAsts     = astsForStatement(stmt.getBody)
+    val code         = s"do {...} while (${stmt.getCondition.toString})"
+    val lineNumber   = line(stmt)
+    val columnNumber = column(stmt)
+
+    doWhileAst(conditionAst, stmtAsts, Some(code), lineNumber, columnNumber)
+  }
+
+  private[statements] def astForWhile(stmt: WhileStmt): Ast = {
+    val conditionAst = astsForExpression(stmt.getCondition, ExpectedType.Boolean).headOption
+    val stmtAsts     = astsForStatement(stmt.getBody)
+    val code         = s"while (${stmt.getCondition.toString})"
+    val lineNumber   = line(stmt)
+    val columnNumber = column(stmt)
+
+    whileAst(conditionAst, stmtAsts, Some(code), lineNumber, columnNumber)
+  }
+
+  private[statements] def astForIf(stmt: IfStmt): Ast = {
+    val ifNode =
+      NewControlStructure()
+        .controlStructureType(ControlStructureTypes.IF)
+        .lineNumber(line(stmt))
+        .columnNumber(column(stmt))
+        .code(s"if (${stmt.getCondition.toString})")
+
+    val conditionAst =
+      astsForExpression(stmt.getCondition, ExpectedType.Boolean).headOption.toList
+
+    val thenAsts = astsForStatement(stmt.getThenStmt)
+    val elseAst  = astForElse(stmt.getElseStmt.toScala).toList
+
+    val ast = Ast(ifNode)
+      .withChildren(conditionAst)
+      .withChildren(thenAsts)
+      .withChildren(elseAst)
+
+    conditionAst.flatMap(_.root.toList) match {
+      case r :: Nil =>
+        ast.withConditionEdge(ifNode, r)
+      case _ =>
+        ast
+    }
+  }
+
+  private[statements] def astForElse(maybeStmt: Option[Statement]): Option[Ast] = {
+    maybeStmt.map { stmt =>
+      val elseAsts = astsForStatement(stmt)
+
+      val elseNode =
+        NewControlStructure()
+          .controlStructureType(ControlStructureTypes.ELSE)
+          .lineNumber(line(stmt))
+          .columnNumber(column(stmt))
+          .code("else")
+
+      Ast(elseNode).withChildren(elseAsts)
+    }
+  }
+
+  private[statements] def astForSwitchStatement(stmt: SwitchStmt): Ast = {
+    val switchNode =
+      NewControlStructure()
+        .controlStructureType(ControlStructureTypes.SWITCH)
+        .code(s"switch(${stmt.getSelector.toString})")
+
+    val selectorAsts = astsForExpression(stmt.getSelector, ExpectedType.empty)
+    val selectorNode = selectorAsts.head.root.get
+
+    val entryAsts = stmt.getEntries.asScala.flatMap(astForSwitchEntry)
+
+    val switchBodyAst = Ast(NewBlock()).withChildren(entryAsts)
+
+    Ast(switchNode)
+      .withChildren(selectorAsts)
+      .withChild(switchBodyAst)
+      .withConditionEdge(switchNode, selectorNode)
+  }
+
+  private[statements] def astForSynchronizedStatement(stmt: SynchronizedStmt): Ast = {
+    val parentNode =
+      NewBlock()
+        .lineNumber(line(stmt))
+        .columnNumber(column(stmt))
+
+    val modifier = Ast(newModifierNode("SYNCHRONIZED"))
+
+    val exprAsts = astsForExpression(stmt.getExpression, ExpectedType.empty)
+    val bodyAst  = astForBlockStatement(stmt.getBody)
+
+    Ast(parentNode)
+      .withChild(modifier)
+      .withChildren(exprAsts)
+      .withChild(bodyAst)
+  }
+
+  private[statements] def astsForSwitchCases(entry: SwitchEntry): Seq[Ast] = {
+    entry.getLabels.asScala.toList match {
+      case Nil =>
+        val target = NewJumpTarget()
+          .name("default")
+          .code("default")
+        Seq(Ast(target))
+
+      case labels =>
+        labels.flatMap { label =>
+          val jumpTarget = NewJumpTarget()
+            .name("case")
+            .code(label.toString)
+          val labelAsts = astsForExpression(label, ExpectedType.empty).toList
+
+          Ast(jumpTarget) :: labelAsts
+        }
+    }
+  }
+
+  private[statements] def astForSwitchEntry(entry: SwitchEntry): Seq[Ast] = {
+    val labelAsts = astsForSwitchCases(entry)
+
+    val statementAsts = entry.getStatements.asScala.flatMap(astsForStatement)
+
+    labelAsts ++ statementAsts
+  }
+
+  private[statements] def astForReturnNode(ret: ReturnStmt): Ast = {
+    val returnNode = NewReturn()
+      .lineNumber(line(ret))
+      .columnNumber(column(ret))
+      .code(ret.toString)
+    if (ret.getExpression.isPresent) {
+      val expectedType = scope.enclosingMethodReturnType.getOrElse(ExpectedType.empty)
+      val exprAsts     = astsForExpression(ret.getExpression.get(), expectedType)
+      returnAst(returnNode, exprAsts)
+    } else {
+      Ast(returnNode)
+    }
+  }
+
+  private[statements] def astsForLabeledStatement(stmt: LabeledStmt): Seq[Ast] = {
+    val jumpTargetAst = Ast(NewJumpTarget().name(stmt.getLabel.toString))
+    val stmtAst       = astsForStatement(stmt.getStatement).toList
+
+    jumpTargetAst :: stmtAst
+  }
+
+  private[statements] def astForThrow(stmt: ThrowStmt): Ast = {
+    val throwNode = NewCall()
+      .name("<operator>.throw")
+      .methodFullName("<operator>.throw")
+      .lineNumber(line(stmt))
+      .columnNumber(column(stmt))
+      .code(stmt.toString())
+      .dispatchType(DispatchTypes.STATIC_DISPATCH)
+
+    val args = astsForExpression(stmt.getExpression, ExpectedType.empty)
+
+    callAst(throwNode, args)
+  }
+
+  private[statements] def astForCatchClause(catchClause: CatchClause): Ast = {
+    astForBlockStatement(catchClause.getBody)
+  }
+
+  private[statements] def astsForTry(stmt: TryStmt): Seq[Ast] = {
+    val tryNode = NewControlStructure()
+      .controlStructureType(ControlStructureTypes.TRY)
+      .code("try")
+      .lineNumber(line(stmt))
+      .columnNumber(column(stmt))
+
+    val resources = stmt.getResources.asScala.flatMap(astsForExpression(_, expectedType = ExpectedType.empty)).toList
+    val tryAst    = astForBlockStatement(stmt.getTryBlock, codeStr = "try")
+    val catchAsts = stmt.getCatchClauses.asScala.map(astForCatchClause)
+    val catchBlock = Option
+      .when(catchAsts.nonEmpty) {
+        Ast(NewBlock().code("catch")).withChildren(catchAsts)
+      }
+      .toList
+    val finallyAst =
+      stmt.getFinallyBlock.toScala.map(astForBlockStatement(_, "finally")).toList
+
+    val controlStructureAst = Ast(tryNode)
+      .withChild(tryAst)
+      .withChildren(catchBlock)
+      .withChildren(finallyAst)
+
+    resources.appended(controlStructureAst)
+  }
+}
